Adding the changes for the project.
