+++ client/src/App.jsx
@@ -1,34 +1,22 @@
-import { useState } from 'react'
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
-import './App.css'
+import React, {useState} from 'react'
+import Auth from './components/Auth'
 
-function App() {
-  const [count, setCount] = useState(0)
-
+const App = () => {
+ const [token, setToken] = useState(localStorage.getItem("token"));   
   return (
-    <>
-      <div>
-        <a href="https://vitejs.dev" target="_blank">
-          <img src={viteLogo} className="logo" alt="Vite logo" />
-        </a>
-        <a href="https://react.dev" target="_blank">
-          <img src={reactLogo} className="logo react" alt="React logo" />
-        </a>
-      </div>
-      <h1>Vite + React</h1>
-      <div className="card">
-        <button onClick={() => setCount((count) => count + 1)}>
-          count is {count}
-        </button>
-        <p>
-          Edit <code>src/App.jsx</code> and save to test HMR
-        </p>
-      </div>
-      <p className="read-the-docs">
-        Click on the Vite and React logos to learn more
-      </p>
-    </>
+    <div className='App'>
+       {!token ?(
+        <Auth setToken={setToken}/>
+       ) : (
+        <>
+        <div>
+            <h1>Chat Room:room_name</h1>
+            {/* Adding the Room Component here */}
+        </div>
+        </>
+       )
+    }   
+        </div>
   )
 }
