+++ client/src/components/Home.jsx
@@ -1,0 +1,13 @@
+import React, {useState} from 'react'
+import Auth from './Auth'
+
+const Home = () => {
+ const [token, setToken] = useState(localStorage.getItem)('token');   
+  return (
+    <>
+    
+    </>
+  )
+}
+
+export default Home
